+++ js/lib/flynn/flynnCanvas.js
@@ -33,7 +33,7 @@ Flynn.Config.VectorRender = {
         lineSize: 2.0,
         vertexSize: 2.0,
     },
-}
+};
 
 Flynn.Canvas = Class.extend({
 
@@ -214,7 +214,19 @@ Flynn.Canvas = Class.extend({
             //-----------------------------
             // Vector graphic simulation
             //-----------------------------
-            ctx.vectorStart = function(color, is_world, constrained){
+            ctx.vectorStart = function(color, is_world, constrained, alpha){
+                // Args:
+                //    color:  A color string with any of the following forms:
+                //           '#RRGGBB', '#RRGGBBAA'
+                //    is_world: true if vector is using world (rather than screen) coordinates.
+                //    constrained:  If true, then x/y will be truncated to land on integer 
+                //           pixel locations.
+                //    alpha: Alpha transparency in range 1.0 (solid) to 0.0 (transparent)
+                //
+                //  If alpha is supplied as part of the color string and as a function
+                //  parameter then the two alphas will be multiplied.
+                // 
+                
                 if(typeof(is_world)==='undefined'){
                     this.is_world = false;
                 }
@@ -228,6 +240,17 @@ Flynn.Canvas = Class.extend({
                     this.constrained = constrained;
                 }
 
+                // Extract alpha from color string (if present).
+                // If alpha is supplied in the color string AND as a function
+                // parameter then the two alphas are multiplied.
+                alpha = alpha == undefined ? 1.0 : alpha;
+                var length = color.length;
+                if(length == 9){
+                    // Color is '#RRGGBBAA
+                    alpha = alpha * parseInt(color.substring(7), 16)/255;
+                    color = color.substring(0, 7);
+                }
+                
                 var config;
                 switch(Flynn.mcp.options.vectorMode){
                     case Flynn.VectorMode.PLAIN:
@@ -242,13 +265,14 @@ Flynn.Canvas = Class.extend({
                 }
                 var line_color = Flynn.Util.shadeColor(color, config.lineBrightness)
                 this.vectorVertexColor = Flynn.Util.colorOverdrive(color, config.vertexBrightness);
+                this.vectorVertexAlpha = alpha * 0.7;
                 this.index_vector_vertex = 0;
                 this.lineSize = config.lineSize;
                 this.vertexSize = config.vertexSize;
                 this.graphics.lineStyle(
                     config.lineSize,
                     Flynn.Util.parseColor(line_color, true),
-                    1);
+                    alpha);
             };
 
             ctx.vectorLineToUnconstrained = function(x, y){
@@ -302,7 +326,7 @@ Flynn.Canvas = Class.extend({
                 // Draw the (bright) vector vertex points
                 var offset = this.vertexSize / 2;
                 this.graphics.lineStyle();
-                this.graphics.beginFill(Flynn.Util.parseColor(this.vectorVertexColor, true));
+                this.graphics.beginFill(Flynn.Util.parseColor(this.vectorVertexColor, true), this.vectorVertexAlpha);
                 for(var i=0; i<this.index_vector_vertex; i+=2) {
                     this.graphics.drawRect(
                         this.vector_vertices[i] - offset + 0.5,
@@ -363,19 +387,17 @@ Flynn.Canvas = Class.extend({
                     color, fill_color, is_world);
             };
 
-            ctx.vectorRect = function(x, y, width, height, color, fill_color, is_world){
+            ctx.vectorRect = function(x, y, width, height, color, fill_color, is_world, alpha){
                 
                 if(typeof(fill_color)!=='undefined' && fill_color){
                     this.fillStyle = fill_color;
                     this.fillRect(x, y, width, height);
                 }
-
-                if(typeof(is_world)==='undefined'){
-                    is_world = false;
-                }
+                is_world = is_world == undefined ? false: is_world;
+                alpha = alpha == undefined ? 1.0 : alpha;
 
                 // Draw a rect using vectors
-                this.vectorStart(color, is_world);
+                this.vectorStart(color, is_world, false, alpha);
                 this.vectorMoveTo(x, y);
                 this.vectorLineTo(x+width-1, y);
                 this.vectorLineTo(x+width-1, y+height-1);
@@ -384,13 +406,14 @@ Flynn.Canvas = Class.extend({
                 this.vectorEnd();
             };
 
-            ctx.vectorLine = function(x1, y1, x2, y2, color, is_world){
+            ctx.vectorLine = function(x1, y1, x2, y2, color, is_world, alpha){
                 if(typeof(is_world)==='undefined'){
                     is_world = false;
                 }
+                alpha = alpha == undefined ? 1.0 : alpha;
 
                 // Draw a vector line
-                this.vectorStart(color, is_world);
+                this.vectorStart(color, is_world, false, alpha);
                 this.vectorMoveTo(x1, y1);
                 this.vectorLineTo(x2, y2);
                 this.vectorEnd();
@@ -458,13 +481,13 @@ Flynn.Canvas = Class.extend({
                 opts              = opts              || {};
                 opts.text         = Flynn.Util.defaultText(opts.text, '<TEXT>');
                 opts.scale        = opts.scale        || 1.0;
-                opts.x            = Flynn.Util.defaultArg(opts.x, null);
-                opts.y            = Flynn.Util.defaultArg(opts.y, null);
+                opts.x            = opts.x            || null;
+                opts.y            = opts.y            || null;
                 opts.justify      = opts.justify      || 'left';
                 opts.color        = opts.color        || Flynn.Colors.WHITE;
                 opts.is_world     = opts.is_world     || false;
                 opts.font         = opts.font         || Flynn.Font.Normal;
-                opts.angle        = Flynn.Util.defaultArg(opts.angle, null);
+                opts.angle        = opts.angle        || null;
                 opts.aspect_ratio = opts.aspect_ratio || 1.0;
                 opts.spacing      = opts.spacing      || 1.0;
 
@@ -665,7 +688,7 @@ Flynn.Canvas = Class.extend({
                 opts.is_reversed  = opts.is_reversed  || false;
                 opts.is_world     = opts.is_world     || false;
                 opts.font         = opts.font         || Flynn.Font.Normal;
-                opts.stretch      = Flynn.Util.defaultArg(opts.stretch, null);
+                opts.stretch      = opts.stretch      || null;
 
                 var draw_x, draw_y;
                 var text = String(opts.text);